requirements.txt
selenium
requests
urllib3
pillow
